refactor product details view
