+++ lib/ui/screens/hits/paged_hits_grid.dart
@@ -9,7 +9,7 @@ class PagedHitsGrid extends StatelessWidget {
       : super(key: key);
 
   final PagingController<int, Product> pagingController;
-  final ValueSetter<String>? onHitClick;
+  final Function(String)? onHitClick;
 
   @override
   Widget build(BuildContext context) {
